Move other functionality  to Main blueprint
